# mono-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple API for a card game.",
        "title": "Monolithic Card Game API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/card": {
            "post": {
                "description": "Create and save a new card",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Add a new card",
                "parameters": [
                    {
                        "description": "Card object",
                        "name": "card",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Card"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Card"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/card/{id}": {
            "get": {
                "description": "Retrieve a card by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Get a card by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Card ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Card"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing card",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Update a card",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Card ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Card object",
                        "name": "card",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Card"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Card"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a card by its ID",
                "tags": [
                    "cards"
                ],
                "summary": "Delete a card",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Card ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/store/buy": {
            "post": {
                "description": "Buy cards",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "summary": "Buy cards",
                "parameters": [
                    {
                        "description": "Order",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.StoreOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "true",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "false",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/store/buy/btob": {
            "post": {
                "description": "Buy card B2B",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "summary": "Buy card B2B",
                "parameters": [
                    {
                        "description": "Order",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.StoreOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "true",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "false",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/store/cards_to_sell": {
            "get": {
                "description": "Get cards to sell",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "summary": "Get cards to sell",
                "responses": {
                    "200": {
                        "description": "true",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "false",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/store/cards_to_sell/btob": {
            "get": {
                "description": "Get cards to sell B2B",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "summary": "Get cards to sell B2B",
                "responses": {
                    "200": {
                        "description": "true",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "false",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/store/sell": {
            "post": {
                "description": "Sell card",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "summary": "Sell card",
                "parameters": [
                    {
                        "description": "Order",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.StoreOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "true",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "false",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/store/transaction": {
            "get": {
                "description": "Get transactions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "summary": "Get transactions",
                "responses": {
                    "200": {
                        "description": "true",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "false",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Card": {
            "description": "Represents a card in the game",
            "type": "object",
            "properties": {
                "affinity": {
                    "type": "string"
                },
                "attack": {
                    "type": "number"
                },
                "defence": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "energy": {
                    "type": "number"
                },
                "family": {
                    "type": "string"
                },
                "hp": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "imgUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "smallImgUrl": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "model.StoreOrder": {
            "type": "object",
            "properties": {
                "card_id": {
                    "type": "integer"
                },
                "store_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}